dto long 수정, mapper에 sql 추가

+++ src/main/java/com/team2/docgram/dto/BoardDto.java
@@ -19,10 +19,10 @@ public class BoardDto {
 	private String content;
 	private Timestamp date;
 	private int boardSecurity;
-	private int user;
-	private int file;
-	private int relatedBoard;
-	private int board_hashtagList;
+	private Long user;
+	private Long file;
+	private Long relatedBoard;
+	private Long board_hashtagList;
 	
 
 }
