+++ src/main/java/com/team2/docgram/dto/UserDto.java
@@ -23,6 +23,6 @@ public class UserDto {
 	private String email;
 	private String starmark;
 	private String mytag;
-	private int team;
+	private Long team;
 	
 }
